chore: add config file and import Sage util library
