+++ apps/amp-als/dataset-service-next/.openapi-generator-ignore
@@ -24,5 +24,5 @@
 pom.xml
 README.md
 **/application.properties
-# **/OpenApiGeneratorApplication.java
-# **/OpenApiGeneratorApplicationTests.java
+**/OpenApiGeneratorApplication.java
+**/OpenApiGeneratorApplicationTests.java
