--- apps/amp-als/dataset-service-next/src/main/java/org/sagebionetworks/amp/als/dataset/service/OpenApiGeneratorApplication.java
@@ -1,24 +1,0 @@
-package org.sagebionetworks.amp.als.dataset.service;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.FilterType;
-import org.springframework.context.annotation.FullyQualifiedAnnotationBeanNameGenerator;
-
-@SpringBootApplication(nameGenerator = FullyQualifiedAnnotationBeanNameGenerator.class)
-@ComponentScan(
-  basePackages = {
-    "org.sagebionetworks.amp.als.dataset.service",
-    "org.sagebionetworks.amp.als.dataset.service.api",
-    "org.sagebionetworks.amp.als.dataset.service.configuration",
-  },
-  nameGenerator = FullyQualifiedAnnotationBeanNameGenerator.class
-)
-public class OpenApiGeneratorApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(OpenApiGeneratorApplication.class, args);
-  }
-}
