+++ apps/amp-als/dataset-service-next/src/test/java/org/sagebionetworks/amp/als/dataset/service/DatasetServiceApplicationTests.java
@@ -4,10 +4,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class OpenApiGeneratorApplicationTests {
+class DatasetServiceApplicationTests {
 
-    @Test
-    void contextLoads() {
-    }
-
+  @Test
+  void contextLoads() {}
 }
